COM-1917 move AppNavigation
