+++ src/navigation/Home/styles.ts
@@ -1,7 +1,7 @@
 import { StyleSheet } from 'react-native';
-import { PINK } from '../styles/colors';
-import { FIRA_SANS_BOLD } from '../styles/fonts';
-import { MARGIN, TAB_BAR_HEIGHT, TAB_BAR_LABEL_WIDTH } from '../styles/metrics';
+import { PINK } from '../../styles/colors';
+import { FIRA_SANS_BOLD } from '../../styles/fonts';
+import { MARGIN, TAB_BAR_HEIGHT, TAB_BAR_LABEL_WIDTH } from '../../styles/metrics';
 
 const styles = StyleSheet.create({
   tabBar: {
